+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - 595 Implement `TestcontainersContainer.DisposeAsync` thread safe (rename `TestcontainersState` to `TestcontainersStates`)
 - 604 Do not deny all files in the Docker image tarball when a `.dockerignore` entry ends with `/`
 - 610 Do not deny all files in the Docker image tarball when a `.dockerignore` entry ends with `/*`
+- 632 Execute local database scripts (inside the container) against `localhost`
 
 ## [2.1.0]